Extract runtime context types and impls into agent/runtime/context.rs
